Split the old 'result.Result' object
The 'Result' object structure was used to represent the student mark and result to an assessment. Now ranks are hold into a new 'Rank' and marks into a new 'Mark' object. The 'Mark' object is implemented in the 'result' module. The 'Rank' object is implemented into the new 'ranking' module. The 'table.Table' object has been renamed to 'Ranking' and moved in the new 'ranking' module. The old 'table' module has been deleted.