Fix bug in multirank mark calculation
The commit #63356f introduce a bug in multirank mark calculation. This commit introduced the new Mark class, but because all marks after that where reduced to float in range 0-1 then the process of summing marks for multirank was no more ponderated by the coefficient of each test. This commit fix this behavior.