Fix #87
Previously when performing sum() against a list of Mark where there was only one mark, the mark itself was returned. Thus causing the « erase » behavior observed. Implement a __copy__ method (override the default one) to copy the object not the record (so no identifier and no _sa_instance_state attribute).

Forgot to add an import when fixing #87